research_platform/requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
celery
redis
channels
channels-redis
elasticsearch-dsl
transformers
torch
scikit-learn
networkx
PyPDF2
Pillow
django-extensions
djangorestframework-simplejwt
django-filter
pandas
numpy
matplotlib
seaborn
